完成部分的自动布局

+++ 二手车/QYCustomCarCell.m
@@ -119,7 +119,7 @@
     _imageCountLabel.textColor = [UIColor whiteColor];
     _imageCountLabel.backgroundColor = [UIColor blackColor];
     _imageCountLabel.textAlignment = NSTextAlignmentCenter;
-    _imageCountLabel.text = _imageCountLabel.text = [NSString stringWithFormat:@"%ld/%ld",_index,_headerModel.picUrls.count];
+    _imageCountLabel.text = _imageCountLabel.text = [NSString stringWithFormat:@"%ld/%ld",(long)_index,(unsigned long)_headerModel.picUrls.count];
     
     // 添加手势
     UITapGestureRecognizer *tap = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(singleClick:)];
@@ -149,7 +149,7 @@
 - (void)scrollViewDidEndDecelerating:(UIScrollView *)scrollView {
     _index = scrollView.contentOffset.x / scrollView.frame.size.width + 1;
     //更新lable的值
-    _imageCountLabel.text = _imageCountLabel.text = [NSString stringWithFormat:@"%ld/%ld",_index,_headerModel.picUrls.count];
+    _imageCountLabel.text = _imageCountLabel.text = [NSString stringWithFormat:@"%ld/%ld",(long)_index,(unsigned long)_headerModel.picUrls.count];
 }
